Store the authentication in a file and make code less duplicated
I moved the authentication info to a file to make it ambient context.
The user then has to set it once and use the functions without
specifying tokens and client id every time. I made this change only for
the get functions yet. But it should be easy enough to make the change
for the Post function as well.

What are the functions missing at the moment is exception message that
instructs the user to use the New-AthenticationSettings and
Save-AuthenticationSettings to create and save the settings when
authentication fails.

I  also kept the current tests as is, they do not work for me as they
are Acceptance tests, not Unit tests and are dependent on Stefan's token
and info.